move test to /test/
