+++ src/test/index.test.d.ts
@@ -1,0 +1,1 @@
+export {};
